Reorginize transction barching and processing
gcc 5.4.0 produced buggy code for the previous implementation
with lock-free-like list of transactions. Make it simpler and
use mutex and double-linked list instead.

update transaction test to write from multiple threads.

TODO: profile the code and if the single mutex for the
transaction queue turns out to be a problem just create several
queues.